# docs/api.yaml
openapi: 3.0.3
info:
  title: Business Cash Management
  description: Business Cash Management
  version: 1.0.0
servers:
  - url: 'http://localhost:8080'
paths:
...
  /client/{id}:
    get:
      tags:
        - Utils
      summary: Get identified client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      operationId: getClientById
      responses:
        '200':
          description: The identified client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Client'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
  /client:
    get:
      tags:
        - Utils
      summary: Get client's list
      operationId: getClients
      responses:
        '200':
          description: List of client
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Client'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
    put:
      tags:
        - Utils
      summary: Update/Create client's informations
      description: Update client's informations if id provided, create them otherwise
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/Client'
        required: true
      operationId: createOrUpdateClient
      responses:
        '200':
          description: Updated or created clients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Client'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
components:
  schemas:
    BadRequest:
      description: Bad request exception
      type: object
      properties:
        type:
          type: string
          example: BadRequestException
        message:
          type: string
          example: Bad request
...
    Client:
      type: object
      properties:
        id:
          type: number
        name:
          type: string
        contact:
          type: string